Remove SampleClusterArray in favor of std::vector

+++ src/definitions_cxx.hpp
@@ -635,7 +635,7 @@ enum class MIDITransposeControlMethod : uint8_t {
 };
 constexpr auto kNumMIDITransposeControlMethods = util::to_underlying(MIDITransposeControlMethod::CHORD) + 1;
 
-constexpr int32_t kNumClustersLoadedAhead = 2;
+constexpr size_t kNumClustersLoadedAhead = 2;
 
 enum class InputMonitoringMode : uint8_t {
 	SMART,
