+++ src/deluge/model/sample/sample_recorder.h
@@ -77,7 +77,7 @@ public:
 
 	uint32_t numSamplesExtraToCaptureAtEndSyncingWise;
 
-	int32_t firstUnwrittenClusterIndex = 0;
+	size_t firstUnwrittenClusterIndex = 0;
 
 	// Put things in valid state so if we get destructed before any recording, it's all ok
 	int32_t currentRecordClusterIndex = -1;
